Module snmp::asn1

Structs

BitString

Boolean

Counter
A 32-bit unsigned integer, that begins at 0 and increases up to 4,294,967,295, then wraps back to 0.

Counter64
A counter like Counter32 but 64 bits wide, allowing a value from 0 to 18,446,744,073,709,551,615.

Gauge
A 32-bit unsigned integer, that may have a value from 0 to 4,294,967,295 and may increase or decrease, like a gauge. A minimum and maximum value are associated with the gauge, indicating its normal range.

Integer
capable of holding a value from -2,147,483,648 to +2,147,483,647

Null

ObjectIdentifier

OctetString

RawValue

Sequence

Set

TimeTicks
A 32-bit unsigned integer that indicates the number of hundredths of seconds since some arbitrary start date. Used for timestamping and to compute elapsed time.

UTCTime

ValueKind

Enums

NetworkAddress
An IP address, encoded as a 4-byte octet string.

ValueClass

Constants

CONSTRUCTED

PRIMITIVE

Traits

Value

Type Definitions

Counter32

Gauge32

Integer32

IpAddress

Unsigned32